Add support for game version 3.30
Game version 3.30 introduced breaking changes that have been addressed in this commit.
The game made significant changes to the way NPC and world inventory items are handled which necessitated the introduction of an additional id randomizer hook. Because of changes to the code layout, signature scanning has been temporarily disabled until the stability of the new layout can be validated.

fix #22